+++ clang-tools-extra/clang-tidy/daedalean/DaedaleanTidyModule.cpp
@@ -11,6 +11,9 @@
 #include "../ClangTidyModuleRegistry.h"
 #include "CommaOperatorMustNotBeUsedCheck.h"
 #include "DerivedClassesCheck.h"
+#include "LambdaImplicitCaptureCheck.h"
+#include "SwitchStatementCheck.h"
+#include "TernaryOperatorMustNotBeUsedCheck.h"
 #include "LambdaReturnTypeCheck.h"
 
 using namespace clang::ast_matchers;
@@ -27,6 +30,12 @@ public:
         "daedalean-comma-operator-must-not-be-used");
     CheckFactories.registerCheck<DerivedClassesCheck>(
         "daedalean-derived-classes");
+    CheckFactories.registerCheck<LambdaImplicitCaptureCheck>(
+        "daedalean-lambda-implicit-capture");
+    CheckFactories.registerCheck<SwitchStatementCheck>(
+        "daedalean-switch-statement");
+    CheckFactories.registerCheck<TernaryOperatorMustNotBeUsedCheck>(
+        "daedalean-ternary-operator-must-not-be-used");
     CheckFactories.registerCheck<LambdaReturnTypeCheck>(
         "daedalean-lambda-return-type");
   }