+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -215,6 +215,17 @@ New checks
 
   FIXME: add release notes.
 
+- New :doc:`daedalean-lambda-implicit-capture
+  <clang-tidy/checks/daedalean-lambda-implicit-capture>` check.
+
+  FIXME: add release notes.
+
+- New :doc:`daedalean-switch-statement
+  <clang-tidy/checks/daedalean-switch-statement>` check.
+
+- New :doc:`daedalean-ternary-operator-must-not-be-used
+  <clang-tidy/checks/daedalean-ternary-operator-must-not-be-used>` check.
+
 - New :doc:`daedalean-lambda-return-type
   <clang-tidy/checks/daedalean-lambda-return-type>` check.